my-stream-utils/src/main/java/hzt/stream/utils/MyObjects.java
package hzt.stream.utils;

import java.util.Objects;

public final class MyObjects {

    private MyObjects() {
    }

    @SafeVarargs
    public static <T> void requireAllNonNull(final Class<T> aClass, T... objects) {
        int counter = 1;
        for (T t : objects) {
            Objects.requireNonNull(t, aClass.getSimpleName() + " " + counter + " is null");
            counter++;
        }
    }

    public static void requireAllNonNull(Object... objects) {
        var exception = new IllegalArgumentException("Some objects where null");
        int counter = 1;
        for (Object object : objects) {
            if (object == null) {
                exception.addSuppressed(new NullPointerException("object " + counter + " is null"));
            }
            counter++;
        }
        if (exception.getSuppressed().length > 0) {
            throw exception;
        }
    }
}
